Add files via upload
Here are the SolvForMule files. Please read the README file for your language and follow the directions to it send to your calculator. If you want to see examples and more of an explanation of the program, head over to this website: https://www.ticalc.org/archives/files/fileinfo/470/47092.html